services/offering-service/pyproject.toml
[project]
name = "offering-service"
version = "0.1.0"
description = "Offering Service for TMF Product Catalog"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "common-python",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.uv.sources]
common-python = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["offering"]
